+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## [Unreleased]
+### Fixed
+- This app now works with Django 1.10.
+
 ## [0.1.1] - 2016-07-11
 ### Fixed
 - Fixed some bugs with EXIF orientation handling.